Add default args to TransformationInfo API
 Adds default values to TransformationInfo and adds JvmOverloads.

 Converts crop rect to Float to allow for sub-pixel crops and
 to use NaN as a default value.

 The crop rect dimensions can now be automatically determined
 to match the surface resolution if they are not set or are
 set to TransformationInfo.CROP_NONE. This required some
 clean-up in the Transformations utility file and in the
 ViewfinderTransformation / Viewfinder implementations.

Bug: 319244865
Test: TransformationsTest, ViewfinderTransformationTest
RelNote: "TransformationInfo now has default values for all
args. This will allow Viewfinders to be created without
any TransformationInfo, which will default to a source
rotation of 0, no source mirroring, and no crop rect."

+++ camera/camera-compose/src/main/java/androidx/camera/compose/CameraXViewfinder.kt
@@ -134,10 +134,10 @@ public fun CameraXViewfinder(
                                 sourceRotation = transformInfo.rotationDegrees,
                                 isSourceMirroredHorizontally = transformInfo.isMirroring,
                                 isSourceMirroredVertically = false,
-                                cropRectLeft = transformInfo.cropRect.left,
-                                cropRectTop = transformInfo.cropRect.top,
-                                cropRectRight = transformInfo.cropRect.right,
-                                cropRectBottom = transformInfo.cropRect.bottom
+                                cropRectLeft = transformInfo.cropRect.left.toFloat(),
+                                cropRectTop = transformInfo.cropRect.top.toFloat(),
+                                cropRectRight = transformInfo.cropRect.right.toFloat(),
+                                cropRectBottom = transformInfo.cropRect.bottom.toFloat()
                             )
                         )
                 }
